clearingHouse address and auction hash storage locations brought into the Asset struct in collateral token

+++ src/LienToken.sol
@@ -481,15 +481,9 @@ contract LienToken is ERC721, ILienToken, AuthInitializable, AmountDeriver {
     emit AddLien(
       params.lien.collateralId,
       uint8(params.stack.length),
-      params.stack.length,
+      uint8(params.stack.length),
       newStackSlot
     );
-    //    emit LienStackUpdated(
-    //      params.lien.collateralId,
-    //      uint8(params.stack.length),
-    //      StackAction.ADD,
-    //      uint8(newStack.length)
-    //    );
   }
 
   function _createLien(
@@ -955,12 +949,6 @@ contract LienToken is ERC721, ILienToken, AuthInitializable, AmountDeriver {
         ++i;
       }
     }
-    //    emit LienStackUpdated(
-    //      stack[position].lien.collateralId,
-    //      position,
-    //      StackAction.REMOVE,
-    //      uint8(newStack.length)
-    //    );
 
     emit RemoveLien(stack[position].lien.collateralId, position);
   }
